add additional packages to avoid a fat completion package
