+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/completion/contributor/CamelGroovyReferenceContributor.java
@@ -14,23 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.idea.completion;
+package org.apache.camel.idea.completion.contributor;
 
 import com.intellij.codeInsight.completion.CompletionType;
-import org.apache.camel.idea.completion.extension.CamelPropertyPlaceholderSmartCompletionExtension;
+import com.intellij.psi.PsiFile;
+import org.apache.camel.idea.completion.extension.CamelEndpointSmartCompletionExtension;
 
 import static com.intellij.patterns.PlatformPatterns.psiElement;
 
 /**
- * Plugin to hook into the IDEA completion system, to setup smart completion for Camel property placeholders (eg {{foo}} style)
+ * Plugin to hook into the IDEA Groovy language, to setup Camel smart completion for editing Groovy source code.
  */
-public class CamelPropertyPlaceholderReferenceContributor extends CamelContributor {
+@Deprecated
+public class CamelGroovyReferenceContributor extends CamelContributor {
 
-    public CamelPropertyPlaceholderReferenceContributor() {
-        // add hook for smart completion for {{ }} placeholders
-        addCompletionExtension(new CamelPropertyPlaceholderSmartCompletionExtension());
+    public CamelGroovyReferenceContributor() {
+        addCompletionExtension(new CamelEndpointSmartCompletionExtension(false));
         extend(CompletionType.BASIC,
-                psiElement().notNull(), // this works anywhere
+                psiElement().and(psiElement().inside(PsiFile.class).inFile(matchFileType("groovy"))),
                 new EndpointCompletion(getCamelCompletionExtensions())
         );
     }
