+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/completion/endpoint/CamelSmartCompletionEndpointOptions.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.idea.completion;
+package org.apache.camel.idea.completion.endpoint;
 
 import java.util.ArrayList;
 import java.util.List;
