+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/completion/extension/CamelEndpointSmartCompletionExtension.java
@@ -36,9 +36,9 @@ import org.apache.camel.idea.util.IdeaUtils;
 import org.apache.camel.idea.util.StringUtils;
 import org.jetbrains.annotations.NotNull;
 
-import static org.apache.camel.idea.completion.CamelSmartCompletionEndpointOptions.addSmartCompletionSuggestionsContextPath;
-import static org.apache.camel.idea.completion.CamelSmartCompletionEndpointOptions.addSmartCompletionSuggestionsQueryParameters;
-import static org.apache.camel.idea.completion.CamelSmartCompletionEndpointValue.addSmartCompletionForEndpointValue;
+import static org.apache.camel.idea.completion.endpoint.CamelSmartCompletionEndpointOptions.addSmartCompletionSuggestionsContextPath;
+import static org.apache.camel.idea.completion.endpoint.CamelSmartCompletionEndpointOptions.addSmartCompletionSuggestionsQueryParameters;
+import static org.apache.camel.idea.completion.endpoint.CamelSmartCompletionEndpointValue.addSmartCompletionForEndpointValue;
 import static org.apache.camel.idea.util.IdeaUtils.isCaretAtEndOfLine;
 
 /**
